build: introduce Vite frontend pipeline
Moves web/static/{app.js,index.html} sources to web/src/, renames
input.css to style.css, and replaces the Tailwind CLI invocation with
a Vite build (PostCSS + Tailwind + asset bundling). Output is still
committed to web/static/ so go install keeps working.
