+++ package.json
@@ -4,6 +4,10 @@
   "private": true,
   "devDependencies": {
     "@tailwindcss/typography": "^0.5.0",
-    "tailwindcss": "^3.4.0"
+    "autoprefixer": "^10.4.0",
+    "postcss": "^8.4.0",
+    "postcss-import": "^16.1.0",
+    "tailwindcss": "^3.4.0",
+    "vite": "^5.4.0"
   }
 }
