+++ web/src/index.html
@@ -4,8 +4,7 @@
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>notesview</title>
-  <script type="module" crossorigin src="/static/app.js"></script>
-  <link rel="stylesheet" crossorigin href="/static/style.css">
+  <link rel="stylesheet" href="./style.css">
 </head>
 <body>
   <div id="app">
@@ -17,5 +16,6 @@
     <aside id="sidebar" class="hidden"></aside>
     <main id="content"></main>
   </div>
+  <script type="module" src="./app.js"></script>
 </body>
 </html>
